+++ src/test/java/com/vaadin/ui/upload/tests/UploadIT.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.ui.upload;
+package com.vaadin.ui.upload.tests;
 
 import java.io.BufferedWriter;
 import java.io.File;
